+++ Controller/Payment/DeleteUpo.php
@@ -10,7 +10,7 @@ class DeleteUpo extends Action
     /**
      * @var RedirectUrlBuilder
      */
-    private $redirectUrlBuilder;
+//    private $redirectUrlBuilder;
 
     /**
      * @var ModuleConfig
@@ -41,7 +41,7 @@ class DeleteUpo extends Action
      */
     public function __construct(
         \Magento\Framework\App\Action\Context $context,
-        \Nuvei\Checkout\Model\Redirect\Url $redirectUrlBuilder,
+//        \Nuvei\Checkout\Model\Redirect\Url $redirectUrlBuilder,
         \Nuvei\Checkout\Model\Config $moduleConfig,
         \Magento\Framework\Controller\Result\JsonFactory $jsonResultFactory,
         \Nuvei\Checkout\Model\Request\Factory $requestFactory,
@@ -49,7 +49,7 @@ class DeleteUpo extends Action
     ) {
         parent::__construct($context);
 
-        $this->redirectUrlBuilder   = $redirectUrlBuilder;
+//        $this->redirectUrlBuilder   = $redirectUrlBuilder;
         $this->moduleConfig         = $moduleConfig;
         $this->jsonResultFactory    = $jsonResultFactory;
         $this->requestFactory       = $requestFactory;