+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ```
     * Clean new line and backslash symbol from order billing and shipping addresses.
     * Set different delay time in the DMN logic according the environment.
+    * Added additional check for wrong Order status upgrade.
 ```
 
 # 3.1.1